update documentation, relinting, correcting versioning issues, updated Dockerfiles to actually copy VERSION into image & add pymdown-extensions to be installed, renamed yamllint & prettier files

+++ .travis/build-test-push.sh
@@ -12,16 +12,22 @@ export IMAGE_TAG
 
 get_version () {
   echo -e '\n<<< Getting & setting versioning info >>>\n'
-  SEMVER_BUMP="${SEMVER_BUMP}"
-  if CURRENT_VERSION=$(docker run --rm "${IMAGE_NAME}":"${IMAGE_TAG}" cat VERSION 2> /dev/null); then
-    echo "${CURRENT_VERSION}" > VERSION
-  fi
   # allows for starting semantic versioning & overriding auto-calculated value (set within TravisCI)
   if [[ -n "${SEMVER_OVERRIDE}" ]]; then
     echo "${SEMVER_OVERRIDE}" > VERSION
+    NEXT_VERSION=$(cat VERSION)
+    echo "Version: ${NEXT_VERSION}"
+  else
+    CURRENT_VERSION=$(docker run --entrypoint="" --rm "${IMAGE_NAME}":"${IMAGE_TAG}" cat VERSION 2> /dev/null)
+    echo "${CURRENT_VERSION}" > VERSION
+    if [[ "${TRAVIS_BRANCH}" = master ]]; then
+      NEXT_VERSION=$(docker run --rm -it -v "${PWD}":/app -w /app treeder/bump --filename VERSION "${SEMVER_BUMP}")
+      echo "Version: ${NEXT_VERSION}"
+    else
+      NEXT_VERSION=$(cat VERSION)
+      echo "Version: ${NEXT_VERSION}"
+    fi
   fi
-  NEXT_VERSION=$(docker run --rm -it -v "${PWD}":/app -w /app treeder/bump --filename VERSION "${SEMVER_BUMP}")
-  echo "Version: ${NEXT_VERSION}"
   export NEXT_VERSION
 }
 
@@ -32,7 +38,7 @@ build_images () {
     echo -e "\n<<< Building ${DISTRO} image >>>\n"
     docker build --rm -f Dockerfile."${DISTRO}" -t "${IMAGE_NAME}":"${IMAGE_TAG}"-"${DISTRO}" .
   done
-  if docker image ls | tail -n+2 | awk '{print $2}'| grep '<none>'; then
+  if docker image ls | tail -n+2 | awk '{print $2}'| grep '<none>' &> /dev/null; then
     echo -e '\n<<< Cleaning up dangling images >>>\n'
     docker rmi "$(docker images -f dangling=true -q)" 2>&-
   fi 
@@ -44,7 +50,7 @@ install_prereqs () {
   GOSS_VER=$(curl -s "https://api.github.com/repos/aelsabbahy/goss/releases/latest" | grep '"tag_name":' | sed -E 's/.*"v([^"]+)".*/\1/')
   export GOSS_VER
   curl -sL "https://github.com/aelsabbahy/goss/releases/download/v${GOSS_VER}/goss-linux-amd64" -o "${HOME}/bin/goss"
-  curl -sL "https://github.com/aelsabbahy/goss/releases/download/v${GOSS_VER}/dgoss" -o "$HOME/bin/dgoss"
+  curl -sL "https://github.com/aelsabbahy/goss/releases/download/v${GOSS_VER}/dgoss" -o "${HOME}/bin/dgoss"
   # trivy (vuln scanner)
   TRIVY_VER=$(curl -s "https://api.github.com/repos/aquasecurity/trivy/releases/latest" | grep '"tag_name":' | sed -E 's/.*"v([^"]+)".*/\1/')
   export TRIVY_VER
@@ -53,18 +59,18 @@ install_prereqs () {
   # snyk (vuln scanner)
   SNYK_VER=$(curl -s "https://api.github.com/repos/snyk/snyk/releases/latest" | grep '"tag_name":' | sed -E 's/.*"v([^"]+)".*/\1/')
   export SNYK_VER
-  curl -sL "https://github.com/snyk/snyk/releases/download/v$SNYK_VER/snyk-linux" -o "$HOME/bin/snyk"
-  chmod +rx "$HOME"/bin/{goss,dgoss,snyk}
+  curl -sL "https://github.com/snyk/snyk/releases/download/v${SNYK_VER}/snyk-linux" -o "${HOME}/bin/snyk"
+  chmod +rx "${HOME}"/bin/{goss,dgoss,snyk}
 }
 
 vulnerability_scanner () {
-  echo -e '\n<<< Checking image for vulnerabilities >>>\n'
   trivy --clear-cache
   for IMAGE in $(docker image ls | tail -n+2 | awk '{OFS=":";} {print $1,$2}'| grep "${DOCKER_USER}"); do
+    echo -e "\n<<< Checking ${IMAGE} for vulnerabilities >>>\n"
     trivy --exit-code 0 --severity "UNKNOWN,LOW,MEDIUM,HIGH" --light -q "${IMAGE}"
-    echo -e '\n<<< Critical Vulnerabilities >>>\n'
+    echo -e "\n<<< Checking ${IMAGE} for critical vulnerabilities >>>\n"
     trivy --exit-code 1 --severity CRITICAL --light -q "${IMAGE}"
-    if [[ "${TRAVIS_BRANCH}" = master ]]; then
+    if [[ "${TRAVIS_PULL_REQUEST}" = false ]] && [[ "${TRAVIS_BRANCH}" = master ]]; then
       snyk auth "${SNYK_TOKEN}" &> /dev/null
       snyk monitor --docker "${IMAGE_NAME}":"${IMAGE_TAG}" --file=Dockerfile
       for DISTRO in $(find . -type f -iname "Dockerfile.*" -print | cut -d'/' -f2 | cut -d'.' -f 2); do
@@ -106,7 +112,9 @@ install_prereqs
 if [[ "${VULNERABILITY_TEST}" = true ]]; then
   vulnerability_scanner
 fi
-test_images
+if [[ "${DGOSS_TEST}" = true ]]; then
+  test_images
+fi
 if [[ "${TRAVIS_PULL_REQUEST}" = false ]] && [[ "${TRAVIS_BRANCH}" =~ ^(dev|master)$ ]]; then
   push_images
 fi
